requirements.txt
Django==3.2.18
graphene>=2.1,<3
graphene-django>=2.1,<3
graphql-core>=2.1,<3
django-filter>=2
django-graphql-jwt==0.3.0
django-graphql-auth
django-image-cropping==1.7
python-slugify==8.0.1
graphene-file-upload==1.3.0
Pillow==9.5.0
gunicorn==20.1.0
psycopg2-binary==2.9.6
easy-thumbnails==2.8.5
celery[redis]==5.2.7
flower==1.2.0
django-celery-results
django-redis==5.2.0